Update HTMLSrcsetParserTests to run testcases individually
HTMLSrcsetParserTests were organized as an array of testcases. If any
failed, it was not possible to connect the failure back to the testcase
that actually failed, because the failure would look like:
  .../parser/html_srcset_parser_test.cc:189: Failure
  Expected equality of these values:
    test.output_density
      Which is: 1
    candidate.Density()
      Which is: 2
  [  FAILED  ] HTMLSrcsetParserTest.Basic (119 ms)

This failure could come from any of 82 tests.

This patch updates the tests to run parameterized, so test failures
print which test they came from:
 .../parser/html_srcset_parser_test.cc:58: Failure
  Expected equality of these values:
    test.output_density
      Which is: 1
    candidate.Density()
      Which is: 2
  [  FAILED  ] HTMLSrcsetParserTest.Basic/1, where GetParam() =
    device_scale_factor: 2; effective_size: 1; src_input: "";
    srcset_input: "data:,a 2w, data:,b 2x" (69 ms)

Bug: 331925120
Change-Id: Ide173f190e35f52fcfda85e929c3f1157d24d42a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6080036
Auto-Submit: Philip Rogers <pdr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Reviewed-by: Traian Captan <tcaptan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394064}