[Blink] Eliminate redundant check of `canvas2d_bridge_`
This check for whether `canvas2d_bridge_` is non-null occurs after first
checking that there is a ResourceProvider and the context is 2D. If the
first two checks are true, there is guaranteed to be a bridge as well
(reasoning in [1]).

Note: I manually loaded [2] in a local Chrome build with this change and
verified that it passed the conditions described in that file.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/6038064
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/printing/manual/canvas2d-vector-text.html;l=1?q=canvas2d-vector-te&sq=&ss=chromium

Bug: 40280152
Change-Id: I211ebc7e4c1adc1b70af100d0ce6a3b3f16f1ce8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6080333
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394147}