[iOS][QD] Temporary fix for CHECK failure on didSelectRowAtIndexPath
This CL temporary downgrades the severity of the CHECK, so we're able to
investigate why this method is being called on a indexPath other than
ItemIdentifierBrowsingData. This CL also uses a NOTREACHED within if
statments so it's easier to investigate what indexPath is not passing
the CHECK, since the crash reports don't have the output.

Bug: 383066554
Change-Id: Ia4dc7a6652ab50500cd46ccf712f756a8bbb3374
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6080376
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Filipa Senra <fsenra@google.com>
Cr-Commit-Position: refs/heads/main@{#1394199}