Introduce FillingPayload
This will include data needed by FormFiller in order to fill the form.

The CL is a no-op since it currently just sets FillingPayload as an
alias of the currently used variant, but a follow-up CL will add more to
the variant.

Bug: 377228667, 40227071
Change-Id: Ic18a5d326f2014caa5b3cb65b88a5cf02d2c9ddd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6080299
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1394219}