[MPArch GuestView] Create manager to coordinate ZoomControllers for a WebContents.
This CL is a pure refactor of ZoomController to split it into a private
Manager class, stored as WebContentsUserData, and ZoomController.
ZoomController is largely left untouched, except that its lifetime is
now managed by the Manager class. This refactor paves the way to have
multiple ZoomControllers in a single WebContents, which is a requirement
for MPArch GuestView.

Bug: 376084060
Change-Id: I378d99f7a4b2a8806f12fc77deeee126b025effa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6075240
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Reviewed-by: James Maclean <wjmaclean@chromium.org>
Commit-Queue: James Maclean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394275}