[iOS][DBD] Only add time range item if not nil
We're been receiving crashes that we're trying to add a nil item to the
ListModal. This CL fixes that by making sure that the time range item is
not nil before adding it.

We don't know what the root cause is for the pref store being nil, when
the UI is still up and running. However, since this code path is going
to be deprecated soon, we're merging this simpler fix.

Fixed: 40072722
Change-Id: I262ea14f797f53f8b55f1b6bfeafb2765d3578cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6070170
Auto-Submit: Filipa Senra <fsenra@google.com>
Commit-Queue: Filipa Senra <fsenra@google.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394300}