PrivateAggregation: Fix flaky failure due to fenced-frame-network-revocation
When fenced-frame-network-revocation runs first, it reliably breaks two
tests in protected-audience-auction-report-buyers-debug-mode-enabled.

It seems likely that fenced-frame-network-revocation is joining an
interest group and that state persists in subsequent test. Of course,
the protected-audience test has no knowledge of the fenced-frame test;
its expectations assume that browser is not already in the interest
group.

As a quick fix, this CL isolates the fenced-frame-network-revocation
test on the www subdomain.

Bug: 383147019
Change-Id: Id71c73072f0380cad1180bfa16b451e8020a145c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6085735
Commit-Queue: Dan McArdle <dmcardle@chromium.org>
Reviewed-by: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394328}