[ios] Don't show VCN enrollment when switching tab
This Fixes the VCN enrollment sheet crash that occurs when switching
tab. Displaying the sheet is simply omitted in that case. The
rational being that is when you switch tab, you switch context, so
displaying a sheet in that case can be confusing.

Change-Id: I099be9c1c62308a72cd23fbd68abc2326fb392aa
Fixed: 365600761
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6077565
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Slobodan Pejic <slobodan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394341}