[PM] Add OnBeforeFooNodeAdded methods to observers.
The new method is called before nodes are added to the graph, so the
public interface will return null for all outgoing edges. Edges that
are assigned in the constructor, so they're constant for the life of
the node, are exposed to the observer an `pending_foo_node` parameters.

To implement the new semantic. splits the internal
NodeBase::OnJoiningGraph() method into 2 stages:

* OnInitializingProperties sets all properties on the node that don't
  affect the graph topology. It's called before sending
  OnBeforeFooNodeAdded notifations.
* OnInitializingEdges lets the node create connections to other nodes.
  It's called between sending OnBeforeFooNodeAdded notifications and
  OnFooNodeAdded notifications.

Since the original OnFooNodeAdded methods are called after both of
these stages, its behaviour doesn't change, and no existing observers
need updating.

R=pmonette

Bug: 40640034
Change-Id: Idb3d7073b6bdf46164600c37ee45207cc5d3fec2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5477387
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#1394349}