Run git format on `profile_network_context_service.cc`
The file is not formatted properly (in accordance to formatting used in
git cl format). This means when writing code, some auto-formatters
(especially for some that format the entire file vs only the diff) will
create noise unrelated to the interesting changes in an effort to
reformat.

This is an effort to format these files.

Change-Id: Ibcd3aee8f1e23228bce460109d5a4350f6e83393
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5988111
Commit-Queue: Abhi Patel <abhipatel@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394364}