[views-ax] Refactor tooltip text to be cached in AppListItemView
This CL continues the refactor for caching the tooltip text across
Views.

Migration depends on the title's tooltip text and whether new_install_dot_ is visible or not.

From previous CL description and tracking bug info:

The tooltip text is used for accessible descriptions of Views.
As a result, this is needed for the ViewsAX project, this is
because this project involves caching accessible properties of
Views rather than computing them "on the fly", and currently the
tooltip text is not cached across all Views. Some Views do cache
it locally (in those classes) but others just compute it on the fly.

Due to the large number of Views and files involved,
my aim is to complete this refactor iteratively across multiple CLs.
In order for this to be possible, and to keep the View builder
and the compiler happy, the refactor will be split into two separate
parts, the implementation and the cleanup.

For the implementation, we will have to use "clever" naming
so that CLs can be largely independent and we keep compatibility
with the View builder.
As such, for now we will have `SetCachedTooltipText(string)`
and `GetCachedTooltipText()` as the main entryways to get/set this
new cached member (`cahed_tooltip_text_`) on View.
The goal will be to eventually rename these to
`SetTooltipText(string)`, `GetCachedTooltipText()`, and
`tooltip_text_`, or something similar.
However, for now, because some Views have their own version of
`SetTooltipText(string)`, `GetTooltipText()` and the existing
`GetTooltipText(point)`, we will use the `Cached` part in the
naming to keep these separate.
During the refactor, we will make these latter ones simply
"point" towards the new ones that affect the new cached tooltip
text on View.

After all the instances of `GetTooltipText(point)` are refactored
in this manner, I will go back and perform cleanup where it makes
sense. Among possible action items for the clean up will be:

* Rename to just `SetTooltipText(string)`, `GetTooltipText()` and
`tooltip_text_`.
* Remove `GetTooltipText(point)`
* Remove or rename local instances of `GetTooltipText()` i.e.
`Label::GetTooltipText()` etc and if possible use only the
new one on View.

Some views that currently cache their tooltip text to some degree,
such as UserItemButton, will need to still keep a cached
"custom" version of the tooltip text, separate from the one cached
on View. This is due to the fact that these Views have custom
logic for the tooltip text that we would like to keep.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417, 378724151
Change-Id: I12c3a5beb5a36024099a97fe54f7d972be01802e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6054925
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Divyansh Mangal <dmangal@microsoft.com>
Reviewed-by: Sejal Anand <sejalanand@microsoft.com>
Reviewed-by: Vinay Singh <vinaysingh@microsoft.com>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394371}