media/gpu/v4l2: Migrate v4l2_queue to tracking_token
Today a "GenericSharedMemoryID" is used to identify frames in order to
avoid re-mapping frames that have already been mapped to a v4l2_buffer.
This ID either comes from the NativePixmapDmabuf ID, DMA buffer FD, or
GpuMemoryBuffer->GetID().

This replaces the GenericSharedMemoryID with an UnguessableToken, which
comes from the frame's tracking_token if it is set.

For encoders, there is no guarantee of reuse VideoFrame reuse. The image
processor and encoder will ignore the tracking token if it is not set,
and will not attempt to maintain v4l2_buffer affinity.

Bug: b:376526494
Change-Id: Ibf22439f0a43b734e6d9786fd32015c1fe9e3f94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5983448
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394374}