[views-ax] GetAccessibleNodeData clean up
This CL gets rid of some View::GetAccessibleNodeData overrides in unit
tests.

One of the changes included changing the role being set in a test
class since the role was one that shouldn't be set on Views.

Only a few remaining and then we'll get rid of this function on View.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: If52abb8f4d70c06ca44903e583f198f0ade540fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6076764
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1394384}