[messaging] Update public ActivityLog API to include kNone action.
This CL adds support for kNone as an action that could be taken in the
UI for the recent activity page. This simplifies the backend code since
it can always provide a type and is also more future proof if we later
add more types without needing the value on all platforms.

It also updates some of the comments and makes the observer API public.

R=skym@chromium.org

Bug: 380517719
Change-Id: I5df856800289aa4200031b328e72a440f3b3e220
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6081905
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394397}