[reland] [canvas] Use synthetic timer when hidden on mac
When a tab is hidden, the BeginFrameSource is disconnected.  This
causes offscreen canvas to quit drawing.  This CL switches to a
synthetic timer so that some progress is made.

Note that other platforms don't do this, and it's not clear
why.  However, the behavior on mac makes quite a lot of sense;
one would use `kHiddenButPainting` if one wanted to produce
frames on a regular cadance.

Anyway, making a change in canvas seemed like a safer thing to
to do vs changing how viz's frame sources work -- that could have
follow-on consequences for a lot more cases than canvas.

New stuff: PS1 is the most recent reverted version, for
comparison.  Added a bunch of tests that would catch a similar
bug that caused the revert, for some values of "similar".

Bug: 40231455
Change-Id: I597411b686c9d253a7d41372395fdd0b8971afe5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5976582
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394413}