[lensoverlay] Remove WaitForOverlayReadyCheck in browser tests.
This test was disabled a day after the WaitForOverlayReady check was
added so it (seems?) to be related. Also SimulateLeftDragClick is only
used in 2 disabled tests right now so this is safe to remove and see
if it makes a difference in flakiness. FWIW, this WaitForOverlay was
causing the test to crash locally on my Mac.

Change-Id: I21c355edfa355ed2e81b4b865b5f348889c7b454
Bug: 377950420
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6082313
Reviewed-by: Duncan Mercer <mercerd@google.com>
Commit-Queue: Nihar Majmudar <niharm@google.com>
Cr-Commit-Position: refs/heads/main@{#1394414}