[a11y] Syncing Tab's title announcement with focused tab (in JAWS).
Issue: When using JAWS with multiple tabs open, moving between tabs
using the PgUp or PgDown keys causes the screen reader’s
announcements to be out of sync with the currently focused page.

Solution: This bug is due to a regression from CL:5910905,
part of the ViewsAX project, which involved migrating the kname
attribute in the root view. The complexity arose because the
virtual function `GetAccessibleWindowTitle` was overridden by 27
subclasses, making the migration challenging.
One such subclass, browser_view, also overrides
`GetAccessibleWindowTitle`, where the migration depends on the
active tab’s accessible name. Although the migration is complete,
a corner case was missed.

According to the comment here,
https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/views/frame/browser_view.cc;drc=3c4f41da148c72046102157e690e8df13f9a4df9;l=1830
when the active tab changes before its content is loaded, an 
additional accessibility event (kFocusContext) is explicitly thrown.
This event ensures that the screen reader focuses on the
root window and announces the current root window (or Document)
name. If the root view’s accessible name is not updated before
this event is raised, the name and announcement may be out of sync.

To solve this problem, we explicitly update the root view’s name
to the new name before the focus event is fired.
Previously, the root view name was updated through the
`UpdateTitleBar()` function, but the kFocusEvent was triggered
first, causing the out-of-sync issue.

Bug: 382307885
Change-Id: I2e121c4fbb588bec43137758aa6379508370497a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6079881
Reviewed-by: Vinay Singh <vinaysingh@microsoft.com>
Reviewed-by: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394458}