[fedcm] Read and write SD-JWT
This CL introduces a library that manipulates SD-JWTs. It can parse,
and serialize all of the different parts.

It can also Sign and Bind, with interfaces that allow the library to be
tested against the specification and independent of the specific
implementation of the cryptographic algorithms.

The CL that follows this one [1] does the actual cryptographic signing
using BoringSSL's primitives, implementing the interfaces defined
in this CL.

The reader uses base::StringSplitOnce() and the base::Base64UrlDecode()
in the browser process and leaves JSON parsing to the data decoder
helper (which uses Rust for the most part, except on Android).

Once this settles a bit more, we'll re-consider moving the
StringSplitOnce() and the Base64UrlDecode() to Rust too.

The writer can be used in the browser process.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/6016126

[2] Intent to Prototype https://groups.google.com/a/chromium.org/g/blink-dev/c/rwu9wFl0mF4?pli=1

Change-Id: I969c18979b73601be4035e83b5dbe5598cebd278
Bug: 380367784
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5943625
Commit-Queue: Sam Goto <goto@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394460}