[Extensions] Cover more open channel failure metric cases.
Before this change we covered some of the early bail-outs when
attempting to open a channel to an extension, but there are five that
were not. If these cases were hit the metric would be emitted as "hung"
which isn't accurate.

After this we emit for every known control flow of opening a channel to
an extension so any future "hung" metrics are truly because the attempt
hung during one of the stages of opening a channel.

Bug: 371011217
Change-Id: I12eeb62900ef392c0ffc40ac1feb9375c8f95a48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6085803
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Auto-Submit: Justin Lulejian <jlulejian@chromium.org>
Commit-Queue: Justin Lulejian <jlulejian@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394465}