[User Education] More coherent auto-config for V2
Default availability was getting set to 7 days if *any* additional
conditions were specified, but should only happen if additional
conditions specify something can happen a maximum number of times.

Also, adding additional availability if the `used` is changed from the
default didn't make sense in retrospect, so removed that. `used` is
almost always tied to the feature being promoted, so putting a delay on
availability in that case wouldn't be necessary.

This is expected to affect a very small number of IPH (possibly zero)
and at worst will allow IPH that were unnecessarily blocked for a
week after shipping to show immediately. In all likelihood it will
affect zero current IPH, but will help to avoid specific problems
(similar to the one in the upstream CL) in the future.

Bug: 382670027
Change-Id: I57cba63bb5b45a8de8ce4564d4545e9e56b026ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6081755
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394484}