When stopping or disabling a remote audio track, do not disable clones.
Cloned audio tracks in WebRTC share the same track interface as
the original track, which serves as the source of the stream.

When the track interface is disabled through a single audio track,
it affects all tracks that share the same interface.

The solution is to remove this dependency on the shared track interface
when disabling or stopping an individual track.

Bug: 40849402
Change-Id: Ib4c54337cd10797449e856b420396038a45d30da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6052891
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Commit-Queue: Sunggook Chue <sunggch@microsoft.com>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394493}