Wait for sync initialization when sending Sharing messages
Sharing messages can be sent using Vapid or sender id (sync), but there
is a migration to sync-only. During browser startup it's possible that
there are incoming sharing messages even when the sync service is not
initialized yet, which results in using Vapid even when sender id target
could be used.

This CL adds pending messages for the FCM sender to wait until sync is
initialized for the sharing messages which can be sent using sender id.

This is a pre-requisite to remove Vapid support.

Bug: 40253551
Change-Id: I707e561da9404627eebdf601525c8e273dee840d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6082314
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394513}