Count usage of srcset cache optimization
AvoidDownloadIfHigherDensityResourceIsInCache will cause srcset to
select a higher-quality image if it is available in the cache. This
differs from the behavior of image-set, as well as the behavior of
srcset selection in other engines. This patch adds a UseCounter to see
how often this is used.

Bug: 40721826
Change-Id: I5ade4c6e61a3df909e3fa66a4f99ccfcc91893dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6085484
Reviewed-by: Traian Captan <tcaptan@chromium.org>
Auto-Submit: Philip Rogers <pdr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394566}