[PIX] Metrics cleanup - Phase 3
The metrics file can't access the PurchaseActionResult enum because of a
dependency issue making its histogram logging logic convoluted. It is
fixed in 2 CLs. This CL is 1/2.

Move PurchaseActionResult enum to facilitated_payments_utils.h in the
utils directory. This gives the metrics file access to the
PurchaseActionResult enum. It's a no-op.

Low-Coverage-Reason: TRIVIAL_CHANGE This is a cleanup change. Logic-wise, no change is made to the EwalletManager.
Bug: 377126728, 367751320, 379723883
Change-Id: Iff02cc575312319c875c4d0b0020aedadc07dae0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6052675
Commit-Queue: Vishwas Uppoor <vishwasuppoor@google.com>
Reviewed-by: Olivia Saul <jsaul@google.com>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394579}